cambio main y nombre vistas
